refactor: extract shared get_peer_context() builder
Move get_peer_context() + _sanitize_agent_name + prelude templates to
shared/peer_context.py and re-export from peer_inject.py. Add
include_role_marker (default True) and split the bootstrap prelude into a
role-marker line and a charter-crossref line so the marker is conditional.
The default path composes them byte-identically, so SubagentStart output is
unchanged (test_peer_inject corpus passes unmodified).
